+++ package.json
@@ -35,5 +35,5 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://192.168.1.14:3000"
+  "proxy": "http://192.168.1.18:3000"
 }
