+++ src/index.css
@@ -43,3 +43,8 @@
     display: flex;
     justify-content: center;
 }
+
+.navbar-login {
+    display: flex;
+    align-items : center;
+}
